Functional. README updated and converted to markdown.
