+++ Cargo.toml
@@ -10,4 +10,4 @@ crate-type = ["cdylib"]
 
 [dependencies]
 pyo3 = "0.24"
-text2num = "2.6.1"
+text2num = "2.6.2"
